+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## vNext
 
+### Breaking Changes
+- ⚠️ **swok8sdiscovery receiver** : The `DiscoveredBy` relation now has inverted direction.
+
 ## v0.150.4
 - **Added** `swootelentityrefprocessor` for manipulating EntityRefs